Add pdf-rulefind.html
Added the frontend page that will be used for the rulefinder, as
well as modified the backend page for it to serve up the frontend
webpage.  TODO: Add communication between the two or add some sort
of parameters in such that it can tell the backend to give it data.
TODO:  Add the part that will actually process the rulebooks.