Restructure pdf-rulefind
Realized that the pdf rulefinder is going to be complex enough to
merit a dedicated folder.

Began working on a backend file for the pdf rulefinder, but realized
that if I do not want to have to write a rest API from scratch on top of
node's http client, I will probably need to use Express or some other
library.