Add /books path to server
Added a path to show a list of available books.  It currently only
checks whether /user/dmtk/rulebooks exists, but it will eventually
return the list of rulebooks that the DM has made available for
users to peruse.  It will probably return the list in json format
for the sake of modularity.