Made small changes, but came to a realization.
Moved the search method into a fiber, so as to allow long operations
without tying up the main thread.  Realized that I should make the
pdf-to-text translation occur on startup instead of on the fly, so
as to simultaneously make the search method perform well, maintain
decoupled class files, and avoid misusing callbacks.