Partially implement retrieval
Parially implmented the component in which data is retrieved from
Mongo.  Whenever search is called, the book's database entry is now
dumped to the console.

The next step is to either return it usefully or to pass a callback
to the retrieval function, such that it may initiate the search
function.