Determined how I want to handle auth
It will be via retrieving times from the server, encrypting your
username and the times with your password as a key, and sending
that string as a token.  The server can then double check the times
against the current one, as well as the username.  If it fails to
decrypt the data, then the password was wrong.

The server js file has been modified to flesh out the get-times
method, while the client side has been altered to be able to test
it.